moved currency converter and updated dashboard
