+++ .eslintrc
@@ -15,12 +15,14 @@
   },
 
   "rules": {
+      "react/jsx-filename-extension": [ 1, { 
+        "ignoreFilesWithoutCode": true }],
         "jsx-a11y/label-has-for": [ 2, {
             "components": [ "Label" ],
             "required": {
                 "some": [ "nesting", "id" ]
             },
-            "allowChildren": false,
+            "allowChildren": false
         }]
     }
 }
