+++ src/index.js
@@ -1,10 +1,12 @@
-import React, { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
+/* eslint-disable react/jsx-filename-extension */
+import React, { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 
-import App from './App';
-import 'bootstrap/dist/css/bootstrap.min.css';
+import App from "./App";
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./output.css";
 
-const rootElement = document.getElementById('root');
+const rootElement = document.getElementById("root");
 const root = createRoot(rootElement);
 
 root.render(
