+++ src/style.css
@@ -1,13 +1,3 @@
-h1,
-p {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
-    Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-
-h1 {
-  text-align: center;
-}
-
-.container {
-  text-align: center;
-}
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
